refactor: update `fireSingle` to use `Clan` object instead of `UUID` in `ClanProviderImpl`

+++ src/main/java/de/murmelmeister/murmelapi/clan/ClanProviderImpl.java
@@ -89,18 +89,21 @@ public final class ClanProviderImpl implements ClanProvider {
         if (createdAt == null) return null;
 
         Clan clan = new Clan(id, normalizedName, tag, sign, description, ownerId, createdBy, createdAt, null, null);
-        refreshProvider.fireSingle(single, id);
+        refreshProvider.fireSingle(single, clan);
         return clan;
     }
 
     @Override
     public int delete(@NotNull UUID id) {
+        Clan existing = cache.getById(id);
+        if (existing == null) return 0;
+
         @Language("MariaDB")
         String sql = "DELETE FROM %s WHERE id = ?".formatted(TABLE_NAME);
         int row = database.update(sql, stmt -> stmt.setString(1, id.toString()));
         if (row < 1) return 0;
 
-        refreshProvider.fireSingle(single, id);
+        refreshProvider.fireSingle(single, existing);
         return row;
     }
 
@@ -158,7 +161,7 @@ public final class ClanProviderImpl implements ClanProvider {
                 .changedBy(changedBy)
                 .changedAt(changedAt)
                 .build();
-        refreshProvider.fireSingle(single, id);
+        refreshProvider.fireSingle(single, clan);
         return clan;
     }
 }
